+++ src/yaml-utilities.ts
@@ -64,8 +64,9 @@ export function isCurrentRecurringItem(task: RecurringTask): boolean {
 
     if (task.recurAfter && (daysSinceTheBeginningOfTime % task.recurAfter === 0)) { return true; };
     if (task.dateAnnual && (dayjs(task.dateAnnual).format("MM-DD") === dayjs().format("MM-DD"))) { return true; }
-    if (task.dateOnce && (dayjs(task.dateOnce).format("YYYY-MM-DD") === dayjs().format("YYYY-MM-DD"))) { return true; }
+    if (task.dateOnce && (dayjs(task.dateOnce).format("YYYY-MM-DD") === dayjs().format("YYYY-MM-DD"))) { return true; }    
     if (task.dayOfWeek && (todayDay.day() === task.dayOfWeek)) { return true; }
+    if (task.dayOfMonth && (todayDay.date() === task.dayOfMonth )) { return true; }
 
     return false;